Add context-aware ingredient cleaning and extract SQL cleaning to TypeScript.
- New shared module `backend/workers/shared/ingredient-cleaning.ts` consolidates all
  cleaning regex constants (PREPARATION_AND_MARKETING_RE, TRAILING_PACKAGING_RE,
  PACKING_MEDIUM_RE, etc.) and exports context-aware dispatchers: recipe/pantry inputs
  get prep-word + optional-phrase stripping; scraper inputs get product-type hoisting
  and packing-medium stripping.
- `ingredient-standardizer.ts` buildPrompt now calls cleanIngredientByContext(name,
  context) instead of the scraper-only preprocessInputName — recipe queue rows now
  have prep words stripped before the AI call.
- `realtime-standardizer.ts` removes duplicated local cleaners; imports
  cleanRecipeIngredientName from the shared module.
- SQL trigger fn_autolink_recipe_ingredient slimmed via migration: replaced the
  fn_clean_product_name / fn_strip_units_from_name dependency (a 6-pass retail
  product stripper) with minimal inline normalization; TypeScript now owns deeper
  context-aware cleaning.
